Improve verification system and streamline interactions
In this commit, we have improved the verification process by preventing a user from sending multiple verification requests. We also made changes to the MemberJoinEvent and MemberLeaveEvent by fetching Member data asynchronously for better performance. Furthermore, interactions with the 'Add Command' in the 'ContextWhitelist' component were streamlined with 'inter.reply' to respond directly to a context menu interaction. Changes were made to 'RemindButton' and 'ReminderVerification' to enhance functionality.